requirements/core.txt
mysql-connector==2.2.9
oauth2client==4.1.3
google-api-core==1.17.0
